# job-portal-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Portal REST API",
    "description": "Comprehensive REST API documentation for the Job Portal system supporting Candidates, Companies, Jobs, Applications, Messaging, and Shortlisting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred."
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "password"
        ],
        "properties": {
          "first_name": { "type": "string", "example": "Jatin" },
          "last_name": { "type": "string", "example": "Verma" },
          "email": { "type": "string", "format": "email", "example": "jatinverma@example.com" },
          "password": { "type": "string", "example": "Jatin@123" },
          "mobile_no": { "type": "string", "example": "+1234567890" },
          "company_name": { "type": "string", "example": "Acme Corp", "description": "Provide this to automatically register as a Company role" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "jatinverma@example.com" },
          "password": { "type": "string", "example": "Jatin@123" }
        }
      },
...
      "ApplicationRequest": {
        "type": "object",
        "required": [
          "vacancy_id"
        ],
        "properties": {
          "vacancy_id": { "type": "string", "format": "uuid", "example": "11111111-1111-1111-1111-111111111114" }
        }
      },
      "ApplicationStatusUpdate": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": { "type": "string", "enum": ["applied", "shortlisted", "contacted", "hired", "rejected"], "example": "shortlisted" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new user (candidate or company)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "User registered successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": { "type": "string", "format": "uuid" },
                        "candidateId": { "type": "string", "format": "uuid", "nullable": true },
                        "companyId": { "type": "string", "format": "uuid", "nullable": true },
                        "first_name": { "type": "string" },
                        "last_name": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string", "example": "candidate" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login, return JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Login successful" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsIn..." },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string", "format": "uuid" },
                            "first_name": { "type": "string" },
                            "last_name": { "type": "string" },
                            "email": { "type": "string" },
                            "role": { "type": "string", "example": "candidate" },
                            "candidateId": { "type": "string", "format": "uuid", "nullable": true },
                            "companyId": { "type": "string", "format": "uuid", "nullable": true }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/candidates/{id}/applications": {
      "get": {
        "tags": ["Candidates"],
        "summary": "Get candidate's job applications",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "7bc31252-bb79-457b-a040-6191e16f39b6"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/applications": {
      "post": {
        "tags": ["Applications"],
        "summary": "Apply for a job (candidate only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Application submitted successfully" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{id}": {
      "get": {
        "tags": ["Applications"],
        "summary": "Get application detail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "11111111-1111-1111-1111-111111111115"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Applications"],
        "summary": "Update application status (company owner only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "11111111-1111-1111-1111-111111111115"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully"
          }
        }
      }
    },
...
  }
}